Fix: Correct help command and blockBreak event errors
This commit addresses two issues that arose from the initial refactoring of the CommandManager:

1.  **Help Command:** The `help` command was failing because it attempted to use array methods on the new `Map` data structure. This has been fixed by converting the `Map` values to an array before sorting and by updating the command lookup logic.

2.  **blockBreak Event:** The `blockBreak` event subscription was failing because the event name is incorrect in the Minecraft API.
    - The event has been corrected to `playerBreakBlock` in `eventManager.js`.
    - The corresponding handler file has been renamed to `playerBreakBlock.js` for consistency.
    - The handler function has been updated to use the correct event properties.
